Replace measure with audit as action in the IMA policy
Initializing IMA with a policy is necessary on upstream kernels to obtain
the digest of a file. However, the measure action causes a significant
performance degradation if the platform has a TPM. Also the appraise action
is not suitable, as IMA would deny access (it would also refuse to run in
permissive mode if secure boot is enabled).

Replace the measure with the audit action, as sending the file digest to
the system log is faster. Unfortunately, due to the fact that the
dont_audit action is not supported by IMA, auditing is done on every
filesystem.

Modify also the diglim_setup.sh script to ignore the INTEGRITY_RULE events
generated by IMA audit, to avoid the system log to grow excessively due to
the loaded IMA policy.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>