Allow mmap() for exec on owned and non reachable temp files
According to the open() man page, the O_TMPFILE open flag allows the
creation of an unnamed temporary regular file. The additional O_EXCL open
flag also gives the guarantee that no other process could access the
created temp file (it is unnamed and cannot be linked).

Since it is the same process creating the file and mmapping it, allow the
mmap() operation, even if the file is still open for writing (normally
considered as a violation due to the risk of arbitrary code execution).

Also, don't generate a digest list for the digest of a page filled with
zeros in diglim_setup.sh, as the LSM will now grant access without
searching for a digest.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>